Update tclParse.c
remove requirement that ( and ) be escaped inside of [command substitution strings and barewords]. If the scanning for the closing paren is INSIDE a [...[...]...] with balancing, then these parens must be part of an arithmetical portion of the expression and cannot be escaped. This makes the behavior the same as expr. One less edge case to worry about.